# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Agendamentos",
    "description": "API RESTful para sistema de agendamentos com controle de disponibilidade",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
...
  "paths": {
...
    "/webhooks": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Registrar um novo webhook",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      },
      "get": {
        "tags": ["Webhooks"],
        "summary": "Listar webhooks do usuário",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de webhooks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Webhook"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "WebhookEventType": {
        "type": "string",
        "enum": [
          "appointment.created",
          "appointment.updated",
          "appointment.cancelled",
          "appointment.completed",
          "appointment.rescheduled",
          "client.created",
          "specialist.created"
        ]
      },
      "WebhookCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Meu Webhook"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://meuapp.com/webhooks/receive"
          },
          "secret": {
            "type": "string",
            "example": "meu_segredo_para_assinatura"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookEventType"
            },
            "example": ["appointment.created", "appointment.cancelled"]
          }
        },
        "required": ["name", "url", "events"]
      },
...
      "Webhook": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookEventType"
            }
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "WebhookResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "webhook": {
            "$ref": "#/components/schemas/Webhook"
          }
        }
      }
    }
  }
}